requirements.txt
Flask
azure-storage-blob
azure-cosmos
werkzeug